Moved representations and request parameters from Marketplace Module
