+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/CategoryRepresentation.java
@@ -7,10 +7,9 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.marketplace.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.vo.VOCategory;
-import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
 import java.util.ArrayList;
