+++ oscm-rest-api-marketplace/src/test/java/org/oscm/rest/marketplace/data/CategoryRepresentationTest.java
@@ -3,6 +3,7 @@ package org.oscm.rest.marketplace.data;
 import org.junit.jupiter.api.Test;
 import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOCategory;
+import org.oscm.rest.common.representation.CategoryRepresentation;
 
 import java.util.ArrayList;
 import java.util.List;
